# docs/frontdesk-api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Front-Desk Operations API
  version: 2.0.0-draft
  description: >
    Domain API for a front-desk voice agent and the front-desk web application,
    reusable across domains through domain packs (healthcare first). Designed
    so that a conversational LLM needs at most three tools: one to find
    availability, one to manage a booking, one to answer general questions
    from approved answers. All business rules live here, never in the MCP
    layer or the prompt.
  contact:
    name: Front-desk platform workstream
servers:
  - url: https://{environment}.example-product.com/api/v1
    variables:
      environment:
        default: beta
...
security:
  - oauth2: []

# -----------------------------------------------------------------------------
# MCP TOOL MAPPING (read by the MCP builder; ignored by the REST server)
# -----------------------------------------------------------------------------
# Three tools. All are thin wrappers over `Agent` operations. The wrapper adds
# nothing but: header injection from call context, an outcome envelope that is
# already the REST response, and stripping of the injected headers from the
# tool's input schema.
...
paths:
...
  /knowledge:
    get:
      tags: [Knowledge]
      operationId: listKnowledge
      summary: Knowledge entries (approved and drafts)
      security: [{ oauth2: [knowledge.write] }]
      parameters:
        - { name: topic, in: query, schema: { type: string } }
        - { name: approved, in: query, schema: { type: boolean } }
      responses:
        '200':
          description: Entries, by topic.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/KnowledgeList' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Knowledge]
      operationId: createKnowledgeEntry
      summary: Add an answer (draft unless approved)
      description: RULE the agent sees only approved entries, from its next question on (every replica).
      security: [{ oauth2: [knowledge.write] }]
      parameters:
        - $ref: '#/components/parameters/ActingUser'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/KnowledgeEntryInput' }
            example:
              topic: parking
              questions: ["is there parking", "ಪಾರ್ಕಿಂಗ್ ಇದೆಯಾ", "पार्किंग है क्या"]
              answers: { en: "Yes. Free parking is in the basement.", kn: "ಹೌದು. ನೆಲಮಾಳಿಗೆಯಲ್ಲಿ ಉಚಿತ ಪಾರ್ಕಿಂಗ್ ಇದೆ." }
              approved: true
      responses:
        '201':
          description: Stored.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/KnowledgeEntry' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /knowledge/{entryId}:
    parameters:
      - { name: entryId, in: path, required: true, schema: { type: string } }
    put:
      tags: [Knowledge]
      operationId: updateKnowledgeEntry
      summary: Replace an answer; the agent sees the change on its next question
      security: [{ oauth2: [knowledge.write] }]
      parameters:
        - $ref: '#/components/parameters/ActingUser'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/KnowledgeEntryInput' }
      responses:
        '200':
          description: Stored.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/KnowledgeEntry' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Knowledge]
      operationId: deleteKnowledgeEntry
      summary: Remove an answer
      security: [{ oauth2: [knowledge.write] }]
      parameters:
        - $ref: '#/components/parameters/ActingUser'
      responses:
        '204': { description: Removed. }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      description: Client credentials per consumer; tenant bound to the credential.
      flows:
        clientCredentials:
          tokenUrl: https://auth.example-product.com/oauth2/token
          scopes:
            agent: Voice agent facade (all /agent operations)
            directory.write: Manage resources, categories, lexicon
            schedule.write: Manage templates and exceptions
            board.write: Update the live board
            bookings.staff: Staff booking operations, impact lists, notifications
            knowledge.write: Manage the knowledge base (approved answers)
            calls.write: Store call summaries
            calls.read: Read call summaries
  parameters:
...
    ActingUser:
      name: X-Acting-User
      in: header
      required: true
      description: Staff identity for audit.
      schema: { type: string, maxLength: 64 }
...
  responses:
    ValidationFailed:
      description: Invalid request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Missing scope.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Not found, or on verified operations the caller number or customer name did not match. Indistinguishable on purpose.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    ClockTime:
      type: string
      pattern: '^([01][0-9]|2[0-3]):[0-5][0-9]$'
      example: "15:30"
...
    Slot:
      type: object
      required: [slotId, kind, available]
      properties:
        slotId: { type: string, example: slot_ses_res_garima_2026-09-26_2_04 }
        kind: { type: string, enum: [SEQUENCE, TIMED] }
        position: { type: integer, minimum: 1, nullable: true, description: SEQUENCE only. }
        start:
          allOf: [{ $ref: '#/components/schemas/ClockTime' }]
          nullable: true
          description: TIMED only.
        end:
          allOf: [{ $ref: '#/components/schemas/ClockTime' }]
          nullable: true
          description: TIMED only.
        expectedWindow:
          type: object
          nullable: true
          description: SEQUENCE only. Server-computed from start and throughput; spoken as "around".
          properties:
            from: { $ref: '#/components/schemas/ClockTime' }
            to: { $ref: '#/components/schemas/ClockTime' }
        available: { type: boolean }
...
    KnowledgeEntryInput:
      type: object
      required: [topic, questions, answers]
      properties:
        topic: { type: string, pattern: '^[a-z0-9_]+$', maxLength: 50 }
        questions:
          type: array
          minItems: 1
          maxItems: 50
          items: { type: string, maxLength: 200 }
          description: Ways callers ask, in any language or script.
        answers:
          type: object
          additionalProperties: { type: string, maxLength: 1000 }
          description: Approved spoken answer per language.
        action: { type: string, enum: [ANSWER, TRANSFER_DESK], default: ANSWER }
        destination: { type: string, nullable: true, description: Required for TRANSFER_DESK; a transfer destination. }
        approved: { type: boolean, default: false }
    KnowledgeEntry:
      allOf:
        - $ref: '#/components/schemas/KnowledgeEntryInput'
        - type: object
          required: [id, source, updatedAt]
          properties:
            id: { type: string }
            source: { type: string, enum: [PROVIDER, DOCUMENT] }
            updatedAt: { type: string, format: date-time }
    KnowledgeList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/KnowledgeEntry' }
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [VALIDATION_FAILED, NOT_FOUND, UNAUTHORIZED, FORBIDDEN, CONFLICT, IDEMPOTENCY_CONFLICT, SLOT_UNAVAILABLE, RATE_LIMITED, INTERNAL, SERVICE_UNAVAILABLE, UPSTREAM_TIMEOUT]
            message: { type: string, description: "Safe to log; never contains another customer's data." }
            details:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  issue: { type: string }
            currentSlots:
              type: array
              description: Present on SLOT_UNAVAILABLE.
              items: { $ref: '#/components/schemas/Slot' }